requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
flask-sqlalchemy>=3.1.1

google-generativeai>=0.7.0
pdfplumber>=0.10.0
openpyxl>=3.1.0
pandas>=2.0.0
python-dateutil>=2.9.0
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0